Fix parameter names missing from Command Panel.
It appears the problem was caused by missing WS_VISIBLE tags in the static control definitions.  However, to get to this point I did a bitwise compare on the memory structures of the generated dialog vs the dialogs created in the resource editor, and I changed quite a few values before getting things to display, so its possibly affected by some of those earlier changes as well.

Resolves Bug #2871