Resolves #FABMAX-38  - EvalContext type not registered.
Problem was that in the Init function we would force the load of the Util extension, but the init function was not reliably called.

Resolved by moving the Extension load to beside where The EvalContext is created, which delays it until absolutely necessary, but ensures it will always happen on time.