Resolves FABMAX-54: Slow perf vs MaxScript
In a long-ago commit, I disabled the caching due to an error:

>> TODO:
>>  - Disabled caching, as our inVals need to be set every time regardless of whether they have changed in Max

From memory, the issue was that the input vals could be modified by the graph, which would give different results following subsequent iterations.  For now we are going to accept this limitation, but further thought to the best behaviour is recommended.